+++ Backend/src/main/java/com/portfolio/backend/repository/SkillsRepo.java
@@ -4,7 +4,7 @@ import com.portfolio.backend.model.Skills;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillsRepo extends JpaRepository<Skills, Integer>{
+public interface SkillsRepo extends JpaRepository<Skills, Long>{
     Optional<Skills> findByName(String name);
     public boolean existsByName(String name);
 }
