+++ Backend/src/main/java/com/portfolio/backend/repository/XpRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface XpRepository extends JpaRepository<WorkXp, Integer>{
+public interface XpRepository extends JpaRepository<WorkXp,Long>{
     public Optional<WorkXp> findByXpName (String xpName);
     public boolean existsByXpName (String xpName);
